Convert various virHash functions to use size_t / uint32
In preparation for conversion over to use the Murmurhash3
algorithm, convert various virHash APIs to use size_t or
uint32 for their return values/parameters, instead of the
variable size 'unsigned long' or 'int' types

+++ src/util/cgroup.c
@@ -1636,9 +1636,9 @@ cleanup:
 }
 
 
-static unsigned long virCgroupPidCode(const void *name)
+static uint32_t virCgroupPidCode(const void *name)
 {
-    return (unsigned long)name;
+    return (uint32_t)(intptr_t)name;
 }
 static bool virCgroupPidEqual(const void *namea, const void *nameb)
 {
